backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
